bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
yoomoney==0.1.0
aiohttp==3.9.3
aiofiles==23.2.1
